refactor: split ui timer renderer
